resumeTailorAPI/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-docx==1.1.0
openai>=1.30.0
python-dotenv==1.0.1
sqlalchemy==2.0.23
